Application/app/requirements.txt
Flask==2.2.2
pymongo==4.3.2
python-dotenv==0.21.0
requests==2.28.1